+++ package.json
@@ -14,7 +14,6 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "^5.0.1"
-
   },
   "devDependencies": {
     "@babel/core": "^7.21.8",
@@ -31,6 +30,7 @@
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0",
-    "vite": "^4.2.0"
+    "vite": "^4.2.0",
+    "whatwg-fetch": "^3.6.2"
   }
 }
