+++ sonar-project.properties
@@ -10,6 +10,6 @@ sonar.test.exclusions=**/*test.js, **/*spec.js, **/*Test.js
 #sonar.projectVersion=1.0
 
 # Path is relative to the sonar-project.properties file. Replace "\" by "/" on Windows.
-sonar.sources=./AuthHom/AuthHom_Pre-Build/
+# sonar.sources=.
 
 sonar.sourceEncoding=UTF-8
