+++ src/Footer.css
@@ -1,0 +1,18 @@
+.app-footer{
+    display: flex;
+    background-color: #0a351c;
+    height: 30vh;
+    padding: 1rem;
+    flex-direction: column;
+    justify-content: space-between;
+    align-items: center;
+    margin: 1.2rem;
+    margin-top: 0rem;
+  }
+  .app-footer p {
+    color: #ffffff;
+    font-size: smaller;
+    font-style: italic;
+    font-weight: 500;
+    text-align: center;
+  }
