+++ src/Hom.css
@@ -1,0 +1,171 @@
+.solicitud-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  background-color: #bec5bc;
+  padding: 2rem 2rem;
+  border-radius: 12px;
+  margin: 1.2rem;
+  height: calc(100% - 45vh);
+  width: 98%;
+  max-width: 2000px;
+  box-sizing: border-box;
+  align-self: center;
+}
+
+.solicitud-title {
+  background-color: #6c9a06;
+  border-radius: inherit;
+  padding: 0.5rem;
+  text-align: center;
+  color: #fff;
+  width: 80;
+  font-size: 12px;
+}
+
+.solicitud-form {
+  display: flex;
+  flex-direction: column;
+  background-color: #bec5bc;
+  padding: 10px;
+  border-radius: 10px;
+  width: 98%;
+}
+
+.solicitud-form h2 {
+  background-color: #0a351c;
+  color: #fff;
+  padding: 0.4%;
+  border-radius: 5px;
+  text-align: center;
+  width: 50%;
+  margin-left: 25%;
+  font-size: 18px;
+}
+
+.solicitud-form form {
+  display: flex;
+  flex-direction: column;
+  font-weight: bold;
+}
+
+.solicitud-form input {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  margin: 0.8rem;
+  margin-left: 0;
+  width: 70%;
+  border: none;
+  background: none;
+  outline: none;
+  border-bottom: 1px solid rgb(0, 0, 0);
+}
+
+.solicitud-form label:first-child,
+.solicitud-form label:nth-child(3) {
+  width: 48%;
+}
+
+.solicitud-form label:last-child,
+.solicitud-form label:nth-child(4) {
+  width: 48%;
+}
+
+.solicitud-fecha {
+  display: flex;
+  justify-content: space-between;
+  margin-top: 10px;
+  margin-left: 40%;
+}
+
+.solicitud-fecha span:last-child {
+  border-style: solid;
+  border-radius: 20%;
+  border-color: rgb(54, 53, 53);
+  border-width: 2px;
+  padding: 0.2rem;
+  margin-top: -0.2rem;
+  margin-left: 0.3rem;
+  font-weight: lighter;
+}
+
+.solicitud-tabla {
+  padding: 10px;
+  margin: 1.2rem;
+  margin-top: 20px;
+}
+
+.solicitud-tabla form {
+  width: inherit;
+  border-collapse: collapse;
+  margin: 1.2rem;
+}
+
+.rellenar-form {
+  display: flex;
+  margin: 3%;
+}
+
+.rellenar-form  input {
+  display: flex;
+  justify-content: space-between;
+  margin: 0.6rem;
+  margin-left: 0;
+  width: 80%;
+  border: none;
+  background: none;
+  outline: none;
+  border-bottom: 1px solid rgb(0, 0, 0);
+}
+
+.rellenar-form label {
+  margin: 0 10px 0 10px;
+  color: #6c9a06;
+  font-weight: bold;
+}
+
+.rellenar-form span {
+  color: red;
+}
+
+.btn-agregar {
+  background-color: #0a351c;
+  color: #fff;
+  border-radius: 10px;
+  padding: 25%;
+  cursor: pointer;
+  font-weight: 550;
+}
+
+.btn-agregar:hover {
+  background-color: #6c9a06;
+}
+
+.separador {
+  width: 95%;
+  border-style: dashed;
+  color: rgb(47, 46, 46);
+}
+
+.solicitud-buttons {
+  display: flex;
+  justify-content: space-between;
+  margin-top: 20px;
+}
+
+.solicitud-buttons button {
+  background-color: #6c9a06;
+  color: #fff;
+  border: none;
+  border-radius: 10px;
+  padding: 10px 20px;
+  cursor: pointer;
+  flex: 1;
+  margin-right: 10px;
+}
+
+.solicitud-buttons button:last-child {
+  margin-right: 0;
+}
