+++ test/FooterC.test.js
@@ -4,7 +4,7 @@
 import "@testing-library/jest-dom";
 // NOTE: jest-dom adds handy assertions to Jest and is recommended, but not required
 import { render, fireEvent, screen } from "@testing-library/react";
-import  FooterC from "../src/components/FooterC";
+import FooterC from "../src/components/FooterC";
 
 describe("Test de renderizacion", () => {
   let componente;
